backend/requirements.txt
fastapi
uvicorn
gunicorn
pandas
openpyxl
python-dotenv
chromadb
sentence-transformers
groq
pydantic
rapidfuzz
pymongo[srv]
python-jose[cryptography]
bcrypt
duckduckgo-search
